+++ pkg/controller/foundationdbcluster/include_instances.go
@@ -1,0 +1,45 @@
+package foundationdbcluster
+
+import (
+	ctx "context"
+	"fmt"
+	"time"
+
+	fdbtypes "github.com/foundationdb/fdb-kubernetes-operator/pkg/apis/apps/v1beta1"
+)
+
+type IncludeInstances struct{}
+
+func (i IncludeInstances) Reconcile(r *ReconcileFoundationDBCluster, context ctx.Context, cluster *fdbtypes.FoundationDBCluster) (bool, error) {
+	adminClient, err := r.AdminClientProvider(cluster, r)
+	if err != nil {
+		return false, err
+	}
+	defer adminClient.Close()
+
+	addresses := make([]string, 0, len(cluster.Spec.PendingRemovals))
+	for _, address := range cluster.Spec.PendingRemovals {
+		addresses = append(addresses, cluster.GetFullAddress(address))
+	}
+
+	if len(addresses) > 0 {
+		r.Recorder.Event(cluster, "Normal", "IncludingInstances", fmt.Sprintf("Including removed processes: %v", addresses))
+	}
+
+	err = adminClient.IncludeInstances(addresses)
+	if err != nil {
+		return false, err
+	}
+
+	if cluster.Spec.PendingRemovals != nil {
+		cluster.Spec.PendingRemovals = nil
+		r.Update(context, cluster)
+		return false, nil
+	}
+
+	return true, nil
+}
+
+func (i IncludeInstances) RequeueAfter() time.Duration {
+	return 0
+}
